+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/trussed-dev/admin-app/compare/0.2.0...HEAD
 
--
+- Add support for string config fields.
 
 ## [0.2.0] 2026-07-22